Extract Stories API rendering from controller to shared concern
